Update patch set 1
Patch Set 1:

In the multivim the ip is a string. From the tests and TOSCA, fixed_ip_address is a list. In my mind, it needs to be transfered. Maybe get the first string of the list as the ip_address if the interfaces are not changed in R2.

Patch-set: 1
Reviewer: Gerrit User 141 <141@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Verified=0