Update patch set 1
Patch Set 1:

Thanks Maopeng.
Yes, fixed_ip_address is a list. Got the 1st string of the list as ip_address should work.
However, it may be a common case that multiple ip addresses are defined in tosca (ipv4, ipv6...). The handling of ip address in vnflcm which always gets the 1st string is somehow a potential concern.

1. Checked multivim, it seems that it does not strictly accept ip as string only.
2. Checked openstack neutron api, it accepts multiple fixed_ip but the ip_address defined inside is normally a string.
3. Checked with Lianhao before doing this patch, that all ip addresses should be sent to openstack finally.

So the question maybe where should we handle the ip_address if the list has several address items? In lcm or in multivim? For me, perhaps it should be better if addressed in multivim.

Patch-set: 1
Reviewer: Gerrit User 3641 <3641@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0